+++ src/app/[locale]/layout.tsx
@@ -1,15 +1,15 @@
-import '@/styles/global.css';
-import { NextIntlClientProvider } from 'next-intl';
 import { unstable_setRequestLocale } from 'next-intl/server';
+import { NextIntlClientProvider } from 'next-intl';
 import { notFound } from 'next/navigation';
-import React from 'react';
+import type { ReactNode } from 'react';
+import '@/styles/global.css';
 
 const locales = ['en'];
 export function generateStaticParams() {
 	return locales.map((locale) => ({ locale }));
 }
 
-export default async function LocaleLayout({ children, params: { locale } }: { children: React.ReactNode, params: { locale: string } }) {
+export default async function LocaleLayout({ children, params: { locale } }: { children: ReactNode, params: { locale: string } }) {
 	let messages;
 	try {
 		messages = (await import(`../../locales/${locale}.json`)).default;
